Estimator

A generic R6 class that wraps around a torch nn module and can be used to fit and predict the model defined in that module.

Methods

Public methods

Estimator$new()

Estimator$fit()

Estimator$fitEpoch()

Estimator$score()

Estimator$finishFit()

Estimator$printProgress()

Estimator$fitWholeTrainingSet()

Estimator$save()

Estimator$predictProba()

Estimator$predict()

Estimator$itemOrDefaults()

Estimator$clone()

Method new()

Creates a new estimator

Usage

Estimator$new(modelType, modelParameters, estimatorSettings)

Arguments

modelType

The torch nn module to use as model

modelParameters

Parameters to initialize the model

estimatorSettings

Parameters required for the estimator fitting

Method fit()

fits the estimator

Usage

Estimator$fit(dataset, testDataset)

Arguments

dataset

a torch dataset to use for model fitting

testDataset

a torch dataset to use for early stopping

Method fitEpoch()

fits estimator for one epoch (one round through the data)

Usage

Estimator$fitEpoch(dataset, batchIndex)

Arguments

dataset

torch dataset to use for fitting

batchIndex

indices of batches

Method score()

calculates loss and auc after training for one epoch

Usage

Estimator$score(dataset, batchIndex)

Arguments

dataset

The torch dataset to use to evaluate loss and auc

batchIndex

Indices of batches in the dataset

Returns

list with average loss and auc in the dataset

Method finishFit()

operations that run when fitting is finished

Usage

Estimator$finishFit(scores, modelStateDict, epoch, learnRates)

Arguments

scores

validation scores

modelStateDict

fitted model parameters

epoch

list of epochs fit

learnRates

learning rate sequence used so far

Method printProgress()

Print out training progress per epoch

Usage

Estimator$printProgress(scores, trainLoss, delta, currentEpoch)

Arguments

scores

scores returned by `self$score`

trainLoss

training loss

delta

how long did the epoch take

currentEpoch

the current epoch number

Method fitWholeTrainingSet()

Fits whole training set on a specific number of epochs

Usage

Estimator$fitWholeTrainingSet(dataset, learnRates = NULL)

Arguments

dataset

torch dataset

learnRates

learnRateSchedule from CV

Method save()

save model and those parameters needed to reconstruct it

Usage

Estimator$save(path, name)

Arguments

path

where to save the model

name

name of file

Returns

the path to saved model

Method predictProba()

predicts and outputs the probabilities

Usage

Estimator$predictProba(dataset)

Arguments

dataset

Torch dataset to create predictions for

Returns

predictions as probabilities

Method predict()

predicts and outputs the class

Usage

Estimator$predict(dataset, threshold = NULL)

Arguments

dataset

A torch dataset to create predictions for

threshold

Which threshold to use for predictions

Returns

The predicted class for the data in the dataset

Method itemOrDefaults()

select item from list, and if it's null sets a default

Usage

Estimator$itemOrDefaults(list, item, default = NULL)

Arguments

list

A list with items

item

Which list item to retrieve

default

The value to return if list doesn't have item

Returns

the list item or default

Method clone()

The objects of this class are cloneable with this method.

Usage

Estimator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.